Refactor pipeline modules into al_pipeline package
